services/web/requirements.txt
Flask==1.1.2
Flask-SQLAlchemy==2.4.3
Flask-HTTPAuth==4.1.0
Flask-Testing==0.8.0
Flask-Mail==0.9.1
SQLAlchemy==1.3.17
gunicorn==20.0.4
psycopg2-binary==2.8.5
itsdangerous==1.1.0
Werkzeug==1.0.1
flask-cors==3.0.8
flask-bcrypt==0.7.1
pyjwt==1.7.1
flask-accept==0.0.6
celery==4.4.5
python-dateutil==2.8.1
python-editor==1.0.4
requests==2.23.0
pyamqp==0.1.0.7
amqp==2.6.0
coverage==5.1.0
oauthlib==3.1.0
email-validator==1.1.1
flower==0.9.4
pydantic==1.5.1
mimesis==4.0.0
click==7.1.2